refactor tests related to minibatches
